* The parts details side view shows the new stock information and the stock log in two tabs. * Removes deprecated fields in the part dialog (quantity, storage, condition)

+++ src/dialogs/removestockdialog.h
@@ -16,8 +16,8 @@ public:
     ~RemoveStockDialog();
     int exec();
     void setPartUnit(const QString & partUnit);
-    void setAvailableStock(int availableStock);
-    int getStockChange() const;
+    void setAvailableStock(double availableStock);
+    double getStockChange() const;
     QString getComment() const;
 private:
     Ui::RemoveStockDialog *ui;
